requirements.txt
Django==5.0
djangorestframework==4.0
psycopg2-binary